rewrite interpreter
this is much classier and more robust
scope inheritance before this commit only went 1 level deep, it was not recursive
now it is recursive, and also cyclic inheritance is no problem, it should work
work as expected

The rule is that within any strongly connected component of the inheritance graph,
there can be no name collisions. Name collisions across different DCSs are fine,
one masks the other in a defined way.

Haven't tested it much yet but the example works and now some more complex
interpreter architecture is possible